+++ Models/Credits.cs
@@ -11,7 +11,7 @@ namespace PersonalFinance.Models
     public string CredTitle { get; set; }
     public string DebName { get; set; }
     public DateTime CredDateTime { get; set; }
-    public int CredValue { get; set; }
+    public float CredValue { get; set; }
     public string CredNote { get; set; }
 
     }
